// sdk/nodejs/s3/enums.ts
// *** WARNING: this file was generated by the Pulumi Terraform Bridge (tfgen) Tool. ***
// *** Do not edit by hand unless you're certain you know what you are doing! ***

export const CannedAcl = {
    Private: "private",
    PublicRead: "public-read",
    PublicReadWrite: "public-read-write",
    AwsExecRead: "aws-exec-read",
    AuthenticatedRead: "authenticated-read",
    BucketOwnerRead: "bucket-owner-read",
    BucketOwnerFullControl: "bucket-owner-full-control",
    LogDeliveryWrite: "log-delivery-write",
} as const;

/**
 * See https://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html#canned-acl
 */
export type CannedAcl = (typeof CannedAcl)[keyof typeof CannedAcl];